feat: Enable multi-arch builds
AMD64 and ARM64 for now.

The Tyk plugin depends on the tykio/tyk-plugin-compiler image which does
not support any other architecture than AMD64 for now.
See https://github.com/TykTechnologies/tyk-build-env/issues/4

ARMv7 was tried but Kong and Tyk do not support it and it may introduce
issues due to its 32-bit architecture that we don't want to have to
support.

Signed-off-by: Felix Kaechele <fkaechel@cisco.com>