Fix bug when invalid spec is stored in DB
Before this fix, an uploaded spec was:
1. validated by apiclarity
2. stored in the DB
3. refs - expanded
4. given to the speculator library
5. validated by the speculator library

If the speculator doesn't validate the expanded spec, then an error is
returned up to the UI, but the spec is already stored in DB at step 2.

Why would the ref-expansion step would create a invalid spec in step 3,
but was considered as valid in step 1 ?

Let's take the following spec:

{
   "swagger":"2.0",
   "info":{
      "version":"",
      "title":"User",
      "description":"Provide Customer login, register, retrieval, as well as card and address retrieval",
      "license":{
         "name":"MIT",
         "url":"http://github.com/gruntjs/grunt/blob/master/LICENSE-MIT"
      }
   },

here, 'info.version' is defined as an empty string. This spec is
valid (step 1.)

When expanded in getExpandedSpec(), the spec is also Marshal'ed to
bytes. During this Marshaling, the 'version.info' field is completly
removed because the go-openapi 'InfoProps' struct tag of the 'Version'
field is 'omitempty".

type InfoProps struct {
	Description    string       `json:"description,omitempty"`
	Title          string       `json:"title,omitempty"`
	TermsOfService string       `json:"termsOfService,omitempty"`
	Contact        *ContactInfo `json:"contact,omitempty"`
	License        *License     `json:"license,omitempty"`
	Version        string       `json:"version,omitempty"`
}

A solution is to store the api in the DB only when valide according to
APIClarity and according to the speculator.